refactor(app): extract layout persisted state and migration to layout-state.ts
Move the persisted-layout schema, migration, and pure session-key/tab helpers
out of context/layout.tsx into a new sibling context/layout-state.ts. The
factory in context/layout.tsx stays runtime-focused; the schema/migration
layer is now plain functions with its own test file.

Moved verbatim: DEFAULT_* width/height constants, right-panel width constants
(DEFAULT_/MIN_/MAX_RIGHT_PANEL_WIDTH), clampRightPanelWidth, SessionTabs /
SessionView / TabHandoff / ReviewDiffStyle types, ensureSessionKey,
createSessionKeyReader, defaultSidePanelTab, legacyOpenShellTabs,
pruneSessionKeys, nextSessionTabsForOpen, sessionPath, normalize* helpers,
createDefaultLayoutState, legacyRightPanelOpened, migrateStoredLayout.

Stayed in context/layout.tsx: AVATAR_COLOR_KEYS / AvatarColorKey /
getAvatarColors (avatar concern; getAvatarColors is dead code preserved
verbatim, queued for a future cleanup slice) and LocalProject (context-public
Project shape used by pages/controllers). The factory imports 17 symbols back
from ./layout-state. MIN_RIGHT_PANEL_WIDTH / MAX_RIGHT_PANEL_WIDTH are
re-exported from context/layout so session-side-panel.tsx's import surface
stays stable.

context/layout.test.ts is renamed to context/layout-state.test.ts with its
import repointed to ./layout-state. All 38 tests pass (full layout slice
guard tests unchanged).

Storage key "layout" (target "layout.v6"), migrate wiring, and persisted
behavior unchanged. No user-visible behavior / DOM / aria / commands changes.

context/layout.tsx 1165 -> 909. context/layout-state.ts: 280 lines.

Refs #1056.
